requirements.txt
absl-py==0.9.0
albumentations==0.4.6
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blessings==1.7
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.5.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
defusedxml==0.6.0
detectron2==0.2+cu102
entrypoints==0.3
filelock==3.0.12
flake8==3.8.3
funcy==1.14
future==0.18.2
fvcore==0.1.1.post20200716
gdown==3.11.1
google-auth==1.19.1
google-auth-oauthlib==0.4.1
gpustat==0.6.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.1.4
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server==1.2.0
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
-e git+https://github.com/ngTienPhat/medico.git@f8cc9dff415f6ace46940f98a5b745a44dafe514#egg=medico&subdirectory=src
mistune==0.8.4
mock==4.0.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.19.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.2
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==1.7.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pycodestyle==2.6.0
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@4d5343c398d75d7ebae34f51a47769084ba3a613
pydot==1.4.1
pyflakes==2.2.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.11.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.7.4
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
yacs==0.1.7
yapf==0.30.0
zipp==3.1.0
